pyproject.toml
[project]
name = "research-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.4.1",
    "ddg>=0.2.2",
    "ddgs>=9.11.3",
    "dotenv>=0.9.9",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "wikipedia>=1.4.0",
]
